+++ tests/testthat/test-struct_equiv.R
@@ -12,7 +12,7 @@ x <- list(
   `static dgCMatrix`= struct_equiv(methods::as(graph, "dgCMatrix")),
   `dynamic array` = struct_equiv(array(graph, dim = c(4,4,3), dimnames = list(rn, rn, 1:3))),
   `dynamic list` = struct_equiv(dyngraph),
-  `dynamic diffnet` = struct_equiv(as_diffnet(dyngraph, c(1,1,3,2)))
+  `dynamic diffnet` = struct_equiv(as_diffnet(dyngraph, c(1L,1L,3L,2L)))
 )
 
 # `Manual` calculations